+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Fixed VRAM leak by rewriting the renderer
+
 ## Abuse 1.0.1 (2025-03-09)
 
 - Corrected the behavior of the icon buttons in the main menu.